Fix reload issues display bug
The reload issues prompt would sometimes display after clicking reload
issues.  This was due to the Gitlab issues tab being set as active as
part of the getIssues() process, but activating a tab does another check
of the API status.  This check would end up occuring prior to
getIssues() finishing.

Also fixed the bug where the Reload issues alert would only show once.
Dismissing the alert with the close button was destroying the object on
the DOM.